requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
firebase-admin==5.0.0
groq==0.3.0
openai==0.27.0
numpy==1.21.2
opencv-python==4.5.3.56
pyzbar==0.1.8
Pillow==8.3.2
requests==2.26.0
python-multipart==0.0.5 